requirements.txt
-i https://pypi.org/simple
dj-database-url==0.5.0
django-heroku==0.3.1
django==2.1.4
djangorestframework==3.9.0
gunicorn==19.9.0
psycopg2==2.7.6.1
pytz==2018.7
whitenoise==4.1.2
